backend/requirements/base.txt
asgiref==3.7.1
Django==3.2.19
django-guardian==2.4.0
django-treebeard==4.7
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
Pillow==9.5.0
pip==23.1.2
psycopg2==2.9.6
PyJWT==1.7.1
pytz==2023.3
setuptools==58.1.0
sqlparse==0.4.4
typing_extensions==4.6.1
django-cors-headers==4.0.0
uritemplate==4.1.1
drf-spectacular==0.26.2
django-filter==23.2
django-auditlog==2.3.0
